handle SIGINT
many ncurses programs such as hexedit use SIGINT (CTRL-C) as the
command to shut down the application, but when one does the same
with vbindiff, the terminal modes are not properly reset, leaving
a messed up terminal that can not even be restored completely with
`reset` command.
so practically everytime i used vbindiff, this happened to me,
accompanied by a "sh... it happened again".