Handle 'panel.h' being in an 'ncurses' directory
Some Linux distributions (such as OpenSUSE) do not keep the ncurses
'panel.h' file in the default include path, but rather in a
subdirectory.

Update the configure script and curses/ConWin.hpp file to look in both
places for a 'panel.h' file. 'panel.h' is still preferred over
'ncurses/panel.h', as we expect other ncurses headers to be in the
include path.